Move TreeInfoView to standalone component
